Segfault sur realloc LARGE resolu

+++ src/ft_putnbrendl.c
@@ -6,13 +6,13 @@
 /*   By: banthony <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/04/30 13:54:11 by banthony          #+#    #+#             */
-/*   Updated: 2017/10/03 14:23:36 by banthony         ###   ########.fr       */
+/*   Updated: 2017/10/04 18:14:05 by banthony         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "malloc.h"
 
-void	ft_putnbrendl(int nb)
+void	ft_putnbrendl(long long nb)
 {
 	ft_putnbr(nb);
 	ft_putchar('\n');
